third_party/blink/web_tests/external/wpt/css/css-conditional/container-queries/style-range-dynamic-expected.txt

This is a testharness.js-based test.
[FAIL] After increasing root line-height
  assert_equals: color of div.test.rlh2 expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
[FAIL] After increasing container font-size for em
  assert_equals: color of div.test.rlh2 expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
[FAIL] After increasing container font-size for lh1
  assert_equals: color of div.test.rlh2 expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
[FAIL] After increasing container line-height
  assert_equals: color of div.test.rlh2 expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
[FAIL] After changing container font-family to TallLineHeight
  assert_equals: color of div.test.rlh2 expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
Harness: the test ran to completion.
